chore: fix example apps
