+++ examples/nextjs/components/auth/WhenLoggedOut.tsx
@@ -1,0 +1,16 @@
+import { useActiveWallet } from '@lens-protocol/react';
+import { ReactNode } from 'react';
+
+export type WhenLoggedOutProps = {
+  children: ReactNode;
+};
+
+export function WhenLoggedOut({ children }: WhenLoggedOutProps) {
+  const { data: wallet, loading } = useActiveWallet();
+
+  if (loading || wallet !== null) {
+    return null;
+  }
+
+  return <>{children}</>;
+}
