+++ examples/nextjs/components/auth/index.tsx
@@ -1,0 +1,3 @@
+export * from './LoginButton';
+export * from './WhenLoggedInWithProfile';
+export * from './WhenLoggedOut';
