+++ examples/nextjs/package.json
@@ -24,6 +24,7 @@
     "next": "^13.1.6",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-hot-toast": "^2.4.0",
     "wagmi": "0.11.7"
   },
   "devDependencies": {
