+++ examples/nextjs/pages/index.tsx
@@ -1,27 +1,8 @@
-import { useActiveProfile, useWalletLogin } from '@lens-protocol/react';
 import Head from 'next/head';
-import { useAccount, useConnect } from 'wagmi';
-import { InjectedConnector } from 'wagmi/connectors/injected';
+
+import { LoginButton, WhenLoggedInWithProfile } from '../components/auth';
 
 export default function Home() {
-  const { execute: login, isPending } = useWalletLogin();
-  const { data: profile } = useActiveProfile();
-
-  const { isDisconnected } = useAccount();
-  const { connectAsync } = useConnect({
-    connector: new InjectedConnector(),
-  });
-  const onLoginClick = async () => {
-    if (isDisconnected) {
-      const { connector } = await connectAsync();
-
-      if (connector instanceof InjectedConnector) {
-        const signer = await connector.getSigner();
-        await login(signer);
-      }
-    }
-  };
-
   return (
     <>
       <Head>
@@ -41,16 +22,10 @@ export default function Home() {
         </p>
       </header>
       <main>
-        {profile && (
-          <p>
-            Welcome <b>@{profile.handle}</b>
-          </p>
-        )}
-        {!profile && (
-          <button disabled={isPending} onClick={onLoginClick}>
-            Log in
-          </button>
-        )}
+        <LoginButton />
+        <WhenLoggedInWithProfile>
+          {({ profile }) => <div>{`Welcome @${profile.handle}`}</div>}
+        </WhenLoggedInWithProfile>
       </main>
     </>
   );
