+++ examples/web-wagmi/src/misc/UseEnabledModules.tsx
@@ -1,7 +1,6 @@
 import { useEnabledModules } from '@lens-protocol/react';
 
-import { LoginButton } from '../components/auth/LoginButton';
-import { WhenLoggedInWithProfile, WhenLoggedOut } from '../components/auth/auth';
+import { LoginButton, WhenLoggedInWithProfile, WhenLoggedOut } from '../components/auth';
 import { ErrorMessage } from '../components/error/ErrorMessage';
 import { Loading } from '../components/loading/Loading';
 import { ModuleCard } from './components/ModuleCard';
